Rework mine block

+++ internal/blockchain/blockchain.go
@@ -7,6 +7,7 @@ package blockchain
 import (
 	"bytes"
 	"fmt"
+	"sync"
 
 	"github.com/nutsdb/nutsdb"
 
@@ -17,15 +18,19 @@ import (
 
 // blockchain implements chain logic.
 type blockchain struct {
-	lastBlock   *types.Block
-	mempool     MemPool
-	genesisHash []byte
 	db          *nutsdb.DB
+	firstBlock  *types.Block
+	lastBlock   *types.Block
+	genesisHash []byte
+	mutex       sync.RWMutex
 }
 
 // New creates a new blockchain instance.
 func New(db *nutsdb.DB) (Blockchain, error) {
-	var lastBlock *types.Block
+	var (
+		lastBlock  *types.Block
+		firstBlock *types.Block
+	)
 
 	if err := db.View(func(tx *nutsdb.Tx) error {
 		iterator := nutsdb.NewIterator(tx, types.BucketBlocks, nutsdb.IteratorOptions{Reverse: true})
@@ -37,34 +42,45 @@ func New(db *nutsdb.DB) (Blockchain, error) {
 
 		lastBlock = types.DeserializeBlock(data)
 
+		if found := iterator.Seek(uint64ToBytes(0)); found {
+			data, err = iterator.Value()
+			if err != nil {
+				return err
+			}
+
+			firstBlock = types.DeserializeBlock(data)
+		}
+
 		return nil
 	}); err != nil {
 		return nil, err
 	}
 
 	return &blockchain{
-		lastBlock: lastBlock,
-		mempool:   newMemPool(),
-		db:        db,
+		lastBlock:  lastBlock,
+		firstBlock: firstBlock,
+		mutex:      sync.RWMutex{},
+		db:         db,
 	}, nil
 }
 
-// AddToMemPool adds a device authentication request to the mem-pool.
-func (b *blockchain) AddToMemPool(request *types.DeviceAuthenticationRequest) {
-	b.mempool.Add(request)
-}
+// // AddToMemPool adds a device authentication request to the mem-pool.
+// func (b *blockchain) AddToMemPool(request *types.DeviceAuthenticationRequest) {
+// 	b.mempool.Add(request)
+// }
 
 // SetGenesisHash sets the genesis block hash.
 func (b *blockchain) SetGenesisHash(hash []byte) {
+	b.mutex.Lock()
+	defer b.mutex.Unlock()
+
 	b.genesisHash = hash
 }
 
 // CreateBlock creates a new block from the mem-pool.
-func (b *blockchain) CreateBlock() (*types.Block, error) {
-	dar := b.mempool.GetFirst()
-	if dar == nil {
-		return nil, ErrEmptyMemPool
-	}
+func (b *blockchain) CreateBlock(dar *types.DeviceAuthenticationRequest) (*types.Block, error) {
+	b.mutex.RLock()
+	defer b.mutex.RUnlock()
 
 	if b.lastBlock != nil {
 		return types.NewBlock(b.lastBlock.Hash, b.lastBlock.Index, dar), nil
@@ -75,15 +91,18 @@ func (b *blockchain) CreateBlock() (*types.Block, error) {
 
 // AddBlock adds a block to the chain.
 func (b *blockchain) AddBlock(block *types.Block) error {
+	b.mutex.Lock()
+	defer b.mutex.Unlock()
+
 	return b.db.Update(func(tx *nutsdb.Tx) error {
 		if b.lastBlock != nil {
-			if block.Index != b.lastBlock.Index+1 {
-				return fmt.Errorf("%w: block %x: index is not valid", ErrBlockValidation, block.Hash)
-			}
-
 			data, _ := tx.Get(types.BucketBlocks, uint64ToBytes(block.Index))
 			if data != nil {
 				return fmt.Errorf("%w: block %x: already exists", ErrBlockValidation, block.Hash)
+			}
+
+			if block.Index != b.lastBlock.Index+1 {
+				return fmt.Errorf("%w: block %x: index is not valid", ErrBlockValidation, block.Hash)
 			}
 
 			if !bytes.Equal(block.PrevHash, b.lastBlock.Hash) {
@@ -97,14 +116,15 @@ func (b *blockchain) AddBlock(block *types.Block) error {
 
 		b.lastBlock = block
 
-		b.mempool.Remove(block.Dar)
-
 		return nil
 	})
 }
 
 // GetBlock returns a block by index.
 func (b *blockchain) GetBlock(index uint64) (*types.Block, error) {
+	b.mutex.RLock()
+	defer b.mutex.RUnlock()
+
 	var block *types.Block
 
 	if err := b.db.View(func(tx *nutsdb.Tx) error {
@@ -125,6 +145,9 @@ func (b *blockchain) GetBlock(index uint64) (*types.Block, error) {
 
 // GetAllBlocks returns all blocks from the chain with pagination.
 func (b *blockchain) GetAllBlocks(from, to uint64) ([]*types.Block, error) {
+	b.mutex.RLock()
+	defer b.mutex.RUnlock()
+
 	var blocks []*types.Block
 
 	if err := b.db.View(func(tx *nutsdb.Tx) error {
@@ -147,8 +170,11 @@ func (b *blockchain) GetAllBlocks(from, to uint64) ([]*types.Block, error) {
 
 // GetLastBlock returns the last block of the chain.
 func (b *blockchain) GetLastBlock() *types.Block {
+	b.mutex.RLock()
+	defer b.mutex.RUnlock()
+
 	if b.lastBlock == nil {
-		return nil
+		return &types.Block{}
 	}
 
 	lastBlock := &types.Block{
@@ -162,8 +188,31 @@ func (b *blockchain) GetLastBlock() *types.Block {
 	return lastBlock
 }
 
+// GetFirstBlock returns the first block of the chain.
+func (b *blockchain) GetFirstBlock() *types.Block {
+	b.mutex.RLock()
+	defer b.mutex.RUnlock()
+
+	if b.firstBlock == nil {
+		return &types.Block{}
+	}
+
+	firstBlock := &types.Block{
+		Hash:      b.firstBlock.Hash,
+		PrevHash:  b.firstBlock.PrevHash,
+		Index:     b.firstBlock.Index,
+		Dar:       b.firstBlock.Dar,
+		Timestamp: b.firstBlock.Timestamp,
+	}
+
+	return firstBlock
+}
+
 // DeleteLastBlock deletes the last block from the chain.
 func (b *blockchain) DeleteLastBlock() error {
+	b.mutex.Lock()
+	defer b.mutex.Unlock()
+
 	return b.db.Update(func(tx *nutsdb.Tx) error {
 		lastIndex := b.lastBlock.Index
 
@@ -180,8 +229,6 @@ func (b *blockchain) DeleteLastBlock() error {
 			}
 
 			b.lastBlock = types.DeserializeBlock(entry.Value)
-
-			b.mempool.Add(b.lastBlock.Dar)
 		}
 
 		return nil
@@ -190,20 +237,24 @@ func (b *blockchain) DeleteLastBlock() error {
 
 // DeleteBlocks delete blocks from the chain.
 func (b *blockchain) DeleteBlocks(from, to uint64) error {
+	b.mutex.Lock()
+	defer b.mutex.Unlock()
+
 	return b.db.Update(func(tx *nutsdb.Tx) error {
-		entries, err := tx.RangeScan(types.BucketBlocks, uint64ToBytes(from), uint64ToBytes(to))
+		entries, err := tx.RangeScan(types.BucketBlocks, uint64ToBytes(from-1), uint64ToBytes(to))
 		if err != nil {
 			return err
 		}
 
-		for _, entry := range entries {
-			block := types.DeserializeBlock(entry.Value)
+		for idx, entry := range entries {
+			if idx == 0 {
+				b.lastBlock = types.DeserializeBlock(entry.Value)
+				continue
+			}
 
 			if err = tx.Delete(types.BucketBlocks, entry.Key); err != nil {
 				return err
 			}
-
-			b.mempool.Add(block.Dar)
 		}
 
 		return nil
